update mini project

+++ ReportEngine/src/main/java/com/miniproject/ReportEngine/Repo/RepoCust.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.miniproject.ReportEngine.Model.listCustomer;
 
-public interface RepoCust extends JpaRepository<listCustomer, Long> {
+public interface RepoCust extends JpaRepository<listCustomer, Integer> {
 
 	@Query(value = "select id, name, address, phone, gender, email, date_of_register, active_account " + 
 			"from list_customer", nativeQuery = true )
